Fix all clippy warnings, add ssh/vnc/logs commands, wire remaining K8s ops
Clippy fixes (25 warnings → 0):
- Replace useless vec![] with array literals in 17 TUI files
- Remove unnecessary u64 casts in cost handler
- Use RangeInclusive::contains in edge module
- Derive Default for RDP types instead of manual impl
- Factor complex type into type alias in NLP search
- Suppress too_many_arguments for api_serve handler

New commands:
- `zorvia ssh <vm>` -- SSH via VM IP or virtctl fallback
- `zorvia vnc <vm>` -- VNC console via virtctl
- `zorvia logs <vm>` -- Stream virt-launcher pod logs via kubectl

Real Kubernetes operations:
- ha-config now patches VM evictionStrategy via K8s API
- evacuate-node cordons node and creates migration CRDs for each VM
- recovery-execute uses SnapshotManager to restore from real snapshots

Backup verification (real, not stub):
- Queries VirtualMachineSnapshot CRD to verify existence, readyToUse
  status, metadata, and snapshot state
- Changed from sync to async with namespace parameter

GFS retention:
- Implemented filter_backups() with full Grandfather-Father-Son retention
  (keep_daily, keep_weekly, keep_monthly, keep_yearly)
- Removed "not yet implemented" warnings

+++ src/handlers/cost.rs
@@ -69,8 +69,8 @@ fn extract_and_calculate_cost(
         name,
         vm_ns,
         cpu,
-        (memory_gi as u64).min(u32::MAX as u64) as u32,
-        (storage_gi as u64).min(u32::MAX as u64) as u32,
+        memory_gi.min(u32::MAX as u64) as u32,
+        storage_gi.min(u32::MAX as u64) as u32,
         period_hours,
     )
 }
